add replay no longer references celery
